Add files via upload
/*To get the given result, we need to add two values of the given vector and compare the result with the target number. To do so, we first need to have a compared value and other comparing values. So first, we will initialize two initializers i and j. i is the compared value, and j is the comparing one.
When i = 0, j = 1. These are indexes of the value in vector nums. We add these two values and compare the result with the target number. Continue through the similar process until the computer finds those that combine and equate the value to the target number. Then I return the result. */
// I know this solution will take much time and space complexity, so I have found some other solutions like Hashmap. I will learn more about this and solve this practice in another way.
/* Calculate the Time Complexity and Space Complexity
Time complexity: O(n)
Space complexity: O(n^2) */
// Add new folder with your name and add your homework in the folder